Fix #51: DockTile broken on Ubuntu 13.04

+++ src/3rdparty/qtdocktile/qtdocktile_unity.cpp
@@ -57,7 +57,7 @@ bool QtDockTile::isAvailable()
 void QtDockTilePrivate::setBadge(int badge)
 {
     QVariantMap map;
-    map.insert(QLatin1String("count"), badge);
+    map.insert(QLatin1String("count"), static_cast<qint64>(badge));
     map.insert(QLatin1String("count-visible"), badge > 0);
     sendMessage(map);
 }
